requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-crispy-forms==1.6.1
Pillow==4.2.1
psycopg2==2.7.2
